feat: About 페이지 - 스토리 데이터 픽스 완료
