+++ src/app/about/components/story/StoryProfile.tsx
@@ -3,8 +3,8 @@
 import { useState } from "react";
 import Image from "next/image";
 import clsx from "clsx";
-import { useResizeObserver } from "../hooks/useResizeObserver";
-import { useStoryRingAnimation } from "../hooks/useStoryRingAnimation";
+import { useResizeObserver } from "../../hooks/useResizeObserver";
+import { useStoryRingAnimation } from "../../hooks/useStoryRingAnimation";
 
 export default function StoryProfile({
   handleOpenStory,
